Handle org switching before pre-loading data
